Start splitting stuff into separate files Makes things easier to keep track of and work with
